Slime - teste case kifera appel a un fonction du game screen

+++ core/src/com/cshv/towerdefense/Mobs/Centaure.java
@@ -158,7 +158,7 @@ public class Centaure extends Mob {
     }
 
     @Override
-    public void addMalus(int malus, int timer) {
+    public void addMalus(float malus, int timer) {
         _malus = malus;
         timerMalus = (new Date().getTime() + timer);
     }
