+++ core/src/com/cshv/towerdefense/Mobs/Griffon.java
@@ -156,7 +156,7 @@ public class Griffon extends Mob {
     }
 
     @Override
-    public void addMalus(int malus, int timer) {
+    public void addMalus(float malus, int timer) {
         _malus = malus;
         timerMalus = (new Date().getTime() + timer);
     }
