+++ core/src/com/cshv/towerdefense/Mobs/Lamia.java
@@ -186,7 +186,7 @@ public class Lamia extends Mob {
     }
 
     @Override
-    public void addMalus(int malus, int timer) {
+    public void addMalus(float malus, int timer) {
         _malus = malus;
         timerMalus = (new Date().getTime() + timer);
     }
