+++ core/src/com/cshv/towerdefense/Mobs/Slime.java
@@ -169,7 +169,7 @@ public class Slime extends Mob {
     }
 
     @Override
-    public void addMalus(int malus, int timer) {
+    public void addMalus(float malus, int timer) {
         _malus = malus;
         timerMalus = (new Date().getTime() + timer);
     }
